monitor-agent/requirements.txt
psutil==5.9.6
httpx==0.25.0
python-dotenv==1.0.0
loguru==0.7.2
apscheduler==3.10.4
